==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Andrea Martinez"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity/",
                      fig.align = "center")
```

# Setting the Environment 

## Set the seed 
```{r set-seed}
set.seed(238428)
```

## Load Libraries 
```{r load-packages}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq,
               install = FALSE)
```

## Load in Data 
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intiution Check 
min(sample_sums(raw_preprocessed_physeq))

metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

#Make n_level a column of characters to enable merge with n_colors 
metadata_df$n_level <- as.character(metadata_df$n_level)

#Make growth_stage a column of characters to enable merge with growth_stage_colors 
metadata_df$growth_stage <- as.character(metadata_df$growth_stage)

head(metadata_df)

# Setting colors for Nitrogen levels
n_colors <- c (
  "0" = "#D9CC3C",
 "150" = "#A0E0BA",
 "300" = "coral2")


growth_stage_colors <- c (
  "30" = "lightblue",
  "75" = "darkblue")

planted_colors <- c (
  "Unplanted" = "springgreen",
  "Planted" = "purple")

```

# Goals

1. Calculate the Hill Diversity of the samples. 
2. Evaluate the rarefaction curves. 
3. Evaluate the Diversity values. 
4. Makes notes of specific samples and their seq depth. 

# Diversity Calculations with iNEXT 

```{r calc-div}
# prepare input data 
iNEXT_input_df <- 
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()
# Quick check
dim(iNEXT_input_df)

# Run iNEXT: Calculate the Hill Numbers 
# Note that: Species in ROWS, Samples in COLUMNS 
# Remember to set the seed! 
#iNEXT_data <- iNEXT(iNEXT_input_df, 
#                   q = c(0,1,2), datatype = "abundance")

#Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the Diversity! 
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
```

# Plot Diversity 
```{r plot-rarefaction}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)

# Make a helper dataframe for plotting with colors 

iNEXT_color_df <- 
  color_df %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(n_colors = n_colors,
            n_level = names(n_colors)),
            by = "n_level")

iNEXT_color_df_2 <- 
  color_df %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(growth_stage_colors = growth_stage_colors,
            growth_stage = names(growth_stage_colors)),
            by = "growth_stage")

iNEXT_color_df_3 <- 
  color_df %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(planted_colors = planted_colors,
            planted = names(planted_colors)),
            by = "planted")
```

# Plot Rarefaction with `ggiNEXT`
```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$n_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$n_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) +
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  theme(legend.position = "none") 

```
Remembering that an Order q of:  

  - 0 = Richness/ Number of Total taxa
  - 1 = Exponential Shannon / Number of "Common" taxa
  - 2 = Inverse Simpson / Number of "Dominant" taxa 
  
*iNEXT will extrapolate the number of sequences to double the sequencing depth of the sample!*

Conclusions from the plot above:  
With increasing weight on abundance, (q = 0,1,2), there is an earlier plateau in the # of ASVs with increasing sequencing depth. This shows that the most highly abundant species are identifiable in early rounds of sequencing due to their high presence. One sample has noticeably higher species richness. 

# Manually plot Diversity 
## Rarefaction
```{r iNEXT-manual, fig.width=6, fig.height=6}
iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # join with metadata 
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(n_colors = n_colors,
                          n_level= names(n_colors)),
            by = "n_level") 
# Inspect 
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

iNEXT_manual_df_2 <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # join with metadata 
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(growth_stage_colors = growth_stage_colors,
                          growth_stage = names(growth_stage_colors)),
            by = "growth_stage") 
# Inspect 
dim(iNEXT_manual_df_2)
str(iNEXT_manual_df_2)


iNEXT_manual_df_3 <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # join with metadata 
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(planted_colors = planted_colors,
                          planted = names(planted_colors)),
            by = "planted") 
# Inspect 
dim(iNEXT_manual_df_3)
str(iNEXT_manual_df_3)

# Plot it - Rarefaction Curve 
iNEXT_manual_df %>%
  # Filter out rows that are calcaulted by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = n_level, group = names)) + 
  geom_line() + 
  # Facet with the n_level to see the samples better 
  facet_grid(Order.q~n_level, scales = "free") + 
  scale_color_manual(values = n_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  #guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```

Abundance-driven community differences seem to be most evident at the medium nitrogen level. There are 2 sample outliers for low ASVs in the Inverse Simpson plot for the low nitrogen treatments. I interpret this to mean these 2 samples have relatively more low-abundance taxa. In the final rarefaction plot presented below, these samples are shown to be from unplanted soils.

A scond rarefaction plot to see rarefaction patterns color coded by treatment, separated by nitrogen treatment.
```{r iNEXT-manual_2, fig.width=6, fig.height=6}
# Plot it - Rarefaction Curve 
iNEXT_manual_df %>%
  # Filter out rows that are calcaulted by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = planted, group = names)) + 
  geom_line() + 
  # Facet with the n_level to see the samples better 
  facet_grid(Order.q~n_level, scales = "free") + 
  scale_color_manual(values = planted_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  #guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```

Lets see how sampling date ("growth stage") rarefaction plots, color coded by nitrogen level give us different information:
```{r iNEXT-manual_3, fig.width=6, fig.height=6}
# Plot it - Rarefaction Curve 
iNEXT_manual_df_2 %>%
  # Filter out rows that are calcaulted by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = n_level, group = names)) + 
  geom_line() + 
  # Facet with the n_level to see the samples better 
  facet_grid(Order.q~growth_stage, scales = "free") + 
  scale_color_manual(values = n_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  #guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```

Rarefaction patterns for high nitrogen samples seem to group together, especially at the earlier time point. One sample outlier in the Inverse Simpson graph for the medium nitrogen treatment at the later time point becomes visible.

Next, lets see rarefaction patterns still color coded by nitrogen treatment, but comparing planted vs unplanted treatments.
```{r iNEXT-manual_4, fig.width=6, fig.height=6}
# Plot it - Rarefaction Curve 
iNEXT_manual_df_3 %>%
  # Filter out rows that are calculted by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = n_level, group = names)) + 
  geom_line() + 
  # Facet with the n_level to see the samples better 
  facet_grid(Order.q~planted, scales = "free") + 
  scale_color_manual(values = n_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  #guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```
Rarefaction patterns for high nitrogen samples still group together in planted and unplanted treatments. The sample outlier at the second time point in the Inverse Simpson plot discussed above appears to be from a planted treatment, with the lowest effective number of ASVs being from a planted, medium nitrogen treatment sample in this plot.

Finally, color code by sampling date and separate by treatment.
```{r iNEXT-manual_5, fig.width=6, fig.height=6}
# Plot it - Rarefaction Curve 
iNEXT_manual_df_3 %>%
  # Filter out rows that are calculted by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot, be sure to group by the samples if using geom_line!
  ggplot(aes(x = m, y= qD, color = growth_stage, group = names)) + 
  geom_line() + 
  # Facet with the n_level to see the samples better 
  facet_grid(Order.q~planted, scales = "free") + 
  scale_color_manual(values = growth_stage_colors) + 
  labs(x = "Number of Sequences (Library Size)", 
       y = "Effective Number of ASVs") + 
  theme_bw() + 
  #guides(color = guide_legend(nrow = 2, byrow = TRUE)) + 
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.title = element_blank()) 
```


Overall richness differs most evidently by time point in the planted treatments, with richness decreasing over time as expected by plant root mediated selection of microbes (via root exudation). There is also a nearly (althouh not entirely) bimodal distribution of richness rarefaction patterns for planted vs unplanted treatments at the medium nitrogen level. Species dominance seems to be influenced by nitrogen treatments and plant presence due to patterns in Inverse Simpson plots.

# Boxplots of Diversity 

Now, let's zoom in on a bit of a better view of the data by visualizing it in a boxplot...  

```{r div-dfs}
n_level_names <- c("Low N","Med N","High N")
names(n_level_names) <- c("0","50","300")

planted_names <- c ("Plant", "No Plant")
names(planted_names) <- c ("Planted", "Unplanted")

growth_stage_names <- c("Day 30","Day 75")
names(growth_stage_names) <- c("30","75")

# Make a dataframe
obs_div_df <- 
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  left_join(data.frame(n_level_names = n_level_names, 
                       n_level = names(n_level_names)), 
                       by = "n_level")
head(obs_div_df)

obs_div_df_2 <-  
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  left_join(data.frame(planted_names = planted_names, 
                      planted = names(planted_names)), 
                       by = "planted")

head(obs_div_df_2)

obs_div_df_3 <-  
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  left_join(data.frame(growth_stage_names = growth_stage_names, 
                      growth_stage = names(growth_stage_names)), 
                       by = "growth_stage")
head(obs_div_df_3)
```


```{r div-boxplot, fig.height = 3.5, fig.width = 9}
# Boxplots by nitrogen level 
# Boxplots by station 
obs_div_df %>%
  ggplot(aes(x = n_level_names, y = qD, fill = n_level, color = n_level)) + 
  facet_wrap(~Order.q, scales = "free") + 
  geom_jitter(size = 2.5) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) + 
  scale_color_manual(values = n_colors) + 
  scale_fill_manual(values = n_colors) + 
  labs(y = "Effective Number of ASVs") + 
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.x = element_blank(),
        legend.title = element_blank())

# Boxplots by plant treatment 
obs_div_df_2 %>%
  ggplot(aes(x = planted_names, y = qD, fill = planted, color = planted)) + 
  facet_wrap(~Order.q, scales = "free") + 
  geom_jitter(size = 2.5) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) + 
  scale_color_manual(values = planted_colors) + 
  scale_fill_manual(values = planted_colors) + 
  labs(y = "Effective Number of ASVs") + 
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.x = element_blank(),
        legend.title = element_blank())

# Boxplots by time point 
obs_div_df_3 %>%
  ggplot(aes(x = growth_stage_names, y = qD, fill = growth_stage, color = growth_stage)) + 
  facet_wrap(~Order.q, scales = "free") + 
  geom_jitter(size = 2.5) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) + 
  scale_color_manual(values = growth_stage_colors) + 
  scale_fill_manual(values = growth_stage_colors) + 
  labs(y = "Effective Number of ASVs") + 
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.x = element_blank(),
        legend.title = element_blank())

```

# Environmental Variable Check 

Applicable environmental variables are not available for this dataset. There may have been more environmental data collected, but the metadata was not available for download. The metadata csv used was constructed by inference for the purpose of this project. I did attempt to find a correlation between available variables, but the output was not informative (below).

```{r environmental-pairs, fig.width=8, fig.height=8}
head(metadata_df)

# Change growth stage and n_level columns back to being numeric
metadata_df$growth_stage <- as.numeric(metadata_df$growth_stage)
metadata_df$n_level <- as.numeric(metadata_df$n_level)

# Pull out environmental variables 
env_df <- 
  metadata_df %>%
  dplyr::select(names, n_level, growth_stage)
# inspect
head(env_df)

# plot the correlations
pairs(dplyr::select(env_df, -names), upper.panel = NULL) 
```

# Diversity vs Nitrogen Addition

```{r div-vs-n_level, fig.height=3.5, fig.width=6}
iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = n_level, y = qD)) + 
  facet_wrap(.~Order.q, scales = "free") + 
  geom_point(aes(color = n_level)) + 
  stat_smooth(method = "lm", formula = y ~poly(x, 2)) + 
  labs(x = "Nitrogen Addition (Kg N ha-1)", y = "# of ASVs") + 
  scale_color_manual(values = n_colors) + 
  theme_bw() + 
  theme(legend.position = "bottom",
        legend.title = element_blank()) 
```

Interpretation: Richness ranges widely for samples in low (0 Kg N ha-1) and medium (150 Kg N ha-1) nitrogen additions and a high nitrogen addition results in intermediate levels of richness. The pattern generally holds when taxa dominance (abundance) is considered in the Shannon and Inverse Simpson plots.

# Session Information
```{r session-info}
devtools::session_info()
```
